+++ CHANGELOG.md
@@ -51,6 +51,8 @@ and this project adheres to
   [#1464](https://github.com/OpenFn/Lightning/issues/1464)
 - Async loading on the history page to improve UX on long DB queries
   [#1279](https://github.com/OpenFn/Lightning/issues/1279)
+- Audit trail events for webhook auth (deletion method) change
+  [#1165](https://github.com/OpenFn/Lightning/issues/1165)
 
 ### Changed